requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-ranged-response==0.2.0
django-recaptcha3==0.4.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
idna==3.4
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
